backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
anthropic
supabase
chromadb
pydantic
websockets
